requirements.txt
aniso8601==9.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==1.16.0
eventlet==0.30.2
Flask==2.1.3
Flask-RESTful==0.3.9
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.2.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
urllib3==1.26.10
Werkzeug==2.2.0
zipp==3.8.1
